Edited model classes for integer only types

+++ lib/models/postava_zadatka.dart
@@ -1,10 +1,10 @@
 class PostavaZadatka {
-  double rata;
-  double max_nabava;
-  double max_kapacitet;
-  double trosak_nabave;
-  double trosak_skladistenja;
-  Map<int, double> razdoblja;
+  int rata;
+  int max_nabava;
+  int max_kapacitet;
+  int trosak_nabave;
+  int trosak_skladistenja;
+  Map<int, int> razdoblja;
 
   PostavaZadatka({required this.rata, required this.max_nabava, required this.max_kapacitet, required this.trosak_nabave, required this.trosak_skladistenja, required this.razdoblja});
 }
